🧹 Add General Commands
- Moved some command to General
