+++ src/Commands/GeneralCommands.ts
@@ -1,0 +1,108 @@
+import { DiscordenoMessage } from 'https://deno.land/x/discordeno@11.2.0/mod.ts';
+import { CommandMap, Command } from '../Interfaces/Command.ts';
+import { GuildModel } from '../Database/index.ts';
+import { SERVER_COMMANDS, USER_COMMANDS } from './index.ts';
+import CONFIG from '../config.ts';
+
+
+/**
+ * Prints Bot Information
+ * @param msg DiscordenoMessage Object
+ * @param cmd Parsed Command Object
+ */
+ async function command_info(msg: DiscordenoMessage, cmd: Command): Promise<any> {
+  // Get Total Guilds Stored
+  const totalGuilds = (await GuildModel.count()) || 0;
+  
+  return msg.send({
+    embed: {
+      title: 'Information',
+      description: `Discord Statistics Bot that monitors user statistics on a server, with the end goal of a personalized User and Server Statistics Experience much like Spotify's Wrapped.
+      Follow Development at: https://github.com/Ciaxur/DiscordStat.bot
+      Support the Developers by donating. Use \`!donate\` for more info.
+
+      This bot is currently in **${totalGuilds} Server${totalGuilds > 1 ? 's' : ''}**`,
+      footer: {
+        text: `Version ${CONFIG.version}`,
+      },
+    },
+  });
+}
+
+/**
+ * Prints help menu for user
+ * @param msg DiscordenoMessage Object
+ * @param cmd Parsed Command Object
+ */
+async function command_help(msg: DiscordenoMessage, cmd: Command): Promise<any> {
+  let initialMessage = `**Announcements**
+    Follow Development at: https://github.com/Ciaxur/DiscordStat.bot
+
+    **General Commands**
+  `;
+
+  return msg.send({
+    embed: {
+      title: 'Help Menu',
+      description: (
+        // General Commands
+        Object.entries(GENERAL_COMMANDS)
+          .reduce((acc, [key, val]) => (acc + `- **${key}**: ${val.description}\n`), initialMessage) +
+        
+        // User Commands
+        Object.entries(USER_COMMANDS)
+          .reduce((acc, [key, val]) => (acc + `- **${key}**: ${val.description}\n`), '\n**User-Specific Commands**\n') +
+
+        // Server Commands
+        Object.entries(SERVER_COMMANDS)
+          .reduce((acc, [key, val]) => (acc + `- **${key}**: ${val.description}\n`), '\n**Server-Specific Commands**\n')
+      ),
+    },
+  });
+}
+
+/**
+ * Retrieves project Version, replying to sender
+ * @param msg DiscordenoMessage Object
+ * @param cmd Parsed Command Object
+ */
+async function command_version(msg: DiscordenoMessage, cmd: Command): Promise<any> {
+  return msg.reply(`Version ${CONFIG.version}`);
+}
+
+/**
+ * Prints information for how to donate to this Bot
+ * @param msg DiscordenoMessage Object
+ * @param cmd Parsed Command Object
+ */
+async function command_donate(msg: DiscordenoMessage, cmd: Command): Promise<any> {
+  msg.send({
+    embed: {
+      title: 'Bot Donation ❤️',
+      image: {
+        url: 'https://ethereum.org/static/a110735dade3f354a46fc2446cd52476/0ee04/eth-home-icon.png',
+      },
+      description: `Donating helps support the development of this bot in the form of future decentralized currency :).
+        - **Ethereum**: 0x1281AD6ce28FD668cf42Ea369ba19413515bD025`
+    },
+  })
+}
+
+export const GENERAL_COMMANDS: CommandMap = {
+  'help': {
+    exec: command_help,
+    description: 'Print the Help Menu',
+  },
+  'info': {
+    exec: command_info,
+    description: 'Prints bot information',
+  },
+  'donate': {
+    exec: command_donate,
+    description: 'Prints Bot instructions for donating',
+  },
+  'version': {
+    exec: command_version,
+    description: 'Prints Bot\'s current version',
+  },
+};
