+++ src/Commands/ServerCommands.ts
@@ -3,7 +3,6 @@ import { CommandMap, Command } from '../Interfaces/Command.ts';
 import { GuildActivityModel, GuildModel } from '../Database/index.ts';
 import { IGuildActivity, IGuild } from '../Interfaces/Database.ts';
 import { GUILD_CACHE } from '../Helpers/Cache.ts';
-import Config from '../config.ts';
 
 /**
  * Prints Command Interaction of users within the server
@@ -43,30 +42,6 @@ async function command_server_interaction(msg: DiscordenoMessage, cmd: Command):
         .reduce((msg, [key, val]) => msg + `- ${key}: ${val}\n`, initialMessage)
     },
   })
-}
-
-/**
- * Prints Bot Information
- * @param msg DiscordenoMessage Object
- * @param cmd Parsed Command Object
- */
-async function command_info(msg: DiscordenoMessage, cmd: Command): Promise<any> {
-  // Get Total Guilds Stored
-  const totalGuilds = (await GuildModel.count()) || 0;
-  
-  return msg.send({
-    embed: {
-      title: 'Information',
-      description: `Discord Statistics Bot that monitors user statistics on a server, with the end goal of a personalized User and Server Statistics Experience much like Spotify's Wrapped.
-      Follow Development at: https://github.com/Ciaxur/DiscordStat.bot
-      Support the Developers by donating. Use \`!donate\` for more info.
-
-      This bot is currently in **${totalGuilds} Server${totalGuilds > 1 ? 's' : ''}**`,
-      footer: {
-        text: `Version ${Config.version}`,
-      },
-    },
-  });
 }
 
 /**
@@ -153,9 +128,5 @@ export const SERVER_COMMANDS: CommandMap = {
   'server-interaction': {
     exec: command_server_interaction,
     description: 'Prints the current server\'s user interaction with this bot',
-  },
-  'info': {
-    exec: command_info,
-    description: 'Prints bot information',
   },
 };
